+++ src/frontend/js/pages/DashboardCreditCardsManagement/DashboardEditCreditCard.spec.tsx
@@ -10,7 +10,10 @@ import {
 import { QueryClientProvider } from '@tanstack/react-query';
 import { IntlProvider } from 'react-intl';
 import fetchMock from 'fetch-mock';
-import { RichieContextFactory as mockRichieContextFactory } from 'utils/test/factories/richie';
+import {
+  UserFactory,
+  RichieContextFactory as mockRichieContextFactory,
+} from 'utils/test/factories/richie';
 import { CreditCardFactory } from 'utils/test/factories/joanie';
 import { DashboardTest } from 'widgets/Dashboard/components/DashboardTest';
 import { SessionProvider } from 'contexts/SessionContext';
@@ -19,13 +22,15 @@ import { expectBreadcrumbsToEqualParts } from 'utils/test/expectBreadcrumbsToEqu
 import { createTestQueryClient } from 'utils/test/createTestQueryClient';
 import { expectBannerError } from 'utils/test/expectBanner';
 import { HttpStatusCode } from 'utils/errors/HttpError';
+import { User } from 'types/User';
+import { OpenEdxApiProfileFactory } from 'utils/test/factories/openEdx';
 
 import { LearnerDashboardPaths } from 'widgets/Dashboard/utils/learnerRoutesPaths';
 
 jest.mock('utils/context', () => ({
   __esModule: true,
   default: mockRichieContextFactory({
-    authentication: { backend: 'fonzie', endpoint: 'https://demo.endpoint' },
+    authentication: { backend: 'fonzie', endpoint: 'https://endpoint.test' },
     joanie_backend: { endpoint: 'https://joanie.endpoint' },
   }).one(),
 }));
@@ -35,7 +40,24 @@ jest.mock('utils/indirection/window', () => ({
 }));
 
 describe('<DahsboardEditCreditCard/>', () => {
+  let richieUser: User;
   beforeEach(() => {
+    richieUser = UserFactory().one();
+    const openEdxProfile = OpenEdxApiProfileFactory({
+      username: richieUser.username,
+      email: richieUser.email,
+      name: richieUser.full_name,
+    }).one();
+    const { 'pref-lang': prefLang, ...openEdxAccount } = openEdxProfile;
+
+    fetchMock.get(
+      `https://endpoint.test/api/user/v1/accounts/${richieUser.username}`,
+      openEdxAccount,
+    );
+    fetchMock.get(`https://endpoint.test/api/user/v1/preferences/${richieUser.username}`, {
+      'pref-lang': prefLang,
+    });
+
     fetchMock.get('https://joanie.endpoint/api/v1.0/orders/', []);
     fetchMock.get('https://joanie.endpoint/api/v1.0/addresses/', []);
   });
@@ -58,7 +80,7 @@ describe('<DahsboardEditCreditCard/>', () => {
 
     await act(async () => {
       render(
-        <QueryClientProvider client={createTestQueryClient({ user: true })}>
+        <QueryClientProvider client={createTestQueryClient({ user: richieUser })}>
           <IntlProvider locale="en">
             <SessionProvider>
               <DashboardTest
@@ -154,7 +176,7 @@ describe('<DahsboardEditCreditCard/>', () => {
 
     await act(async () => {
       render(
-        <QueryClientProvider client={createTestQueryClient({ user: true })}>
+        <QueryClientProvider client={createTestQueryClient({ user: richieUser })}>
           <IntlProvider locale="en">
             <SessionProvider>
               <DashboardTest
@@ -246,7 +268,7 @@ describe('<DahsboardEditCreditCard/>', () => {
 
     await act(async () => {
       render(
-        <QueryClientProvider client={createTestQueryClient({ user: true })}>
+        <QueryClientProvider client={createTestQueryClient({ user: richieUser })}>
           <IntlProvider locale="en">
             <SessionProvider>
               <DashboardTest
@@ -319,7 +341,7 @@ describe('<DahsboardEditCreditCard/>', () => {
 
     await act(async () => {
       render(
-        <QueryClientProvider client={createTestQueryClient({ user: true })}>
+        <QueryClientProvider client={createTestQueryClient({ user: richieUser })}>
           <IntlProvider locale="en">
             <SessionProvider>
               <DashboardTest